+++ src/AD_ReverseDiff.jl
@@ -1,17 +1,12 @@
-import .ReverseDiff, .DiffResults
+#####
+##### Gradient AD implementation using ForwardDiff
+#####
+
+import .ReverseDiff
 
 struct ReverseDiffLogDensity{L, C} <: ADGradientWrapper
     ℓ::L
     gradientconfig::C
-end
-
-Base.show(io::IO, ℓ::ReverseDiffLogDensity) = print(io, "ReverseDiff AD wrapper for ", ℓ.ℓ)
-
-function logdensity(vgb::ValueGradientBuffer, fℓ::ReverseDiffLogDensity, x::AbstractVector)
-    @unpack ℓ, gradientconfig = fℓ
-    result = ReverseDiff.gradient!(DiffResults.MutableDiffResult(vgb),
-                                   x -> logdensity(Real, ℓ, x), x, gradientconfig)
-    ValueGradient(result)
 end
 
 """
@@ -23,4 +18,13 @@ function ADgradient(::Val{:ReverseDiff}, ℓ)
     z = zeros(dimension(ℓ))
     cfg = ReverseDiff.GradientConfig(z)
     ReverseDiffLogDensity(ℓ, cfg)
+end
+
+Base.show(io::IO, ℓ::ReverseDiffLogDensity) = print(io, "ReverseDiff AD wrapper for ", ℓ.ℓ)
+
+function logdensity_and_gradient(fℓ::ReverseDiffLogDensity, x::AbstractVector)
+    @unpack ℓ, gradientconfig = fℓ
+    buffer = _diffresults_buffer(ℓ, x)
+    result = ReverseDiff.gradient!(buffer, x -> logdensity(ℓ, x), x, gradientconfig)
+    _diffresults_extract(result)
 end